requirements.txt
# Core Python utilities
python-dotenv
requests
asyncio
pydantic

# LangChain and LangGraph components
langchain
langgraph
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langgraph-prebuilt
langchain-mcp-adapters

# SQL and DB
sqlalchemy

# Typing extensions (for TypedDict, Annotated)
typing-extensions

fastapi

pytesseract
pdf2image
asyncpg
transformers
pymupdf
nltk
sentence-transformers
uvicorn[standard]
qdrant-client
pydantic